minor changes
apphostcx:
	don't be too verbose when read returns -1

baseproxy:
	small tweak: handle events on cx _after_
	read and write, instead of before.

buffer:
	make exception what() string more concrete

logger:	enclose argument in braces in macros

hostcx:
	testing break in place. Only single read will happen.
	Keep it there and test. Remove in future.

sslcom:
	catch out_of_range exception in parse_peer_hello()

+++ hostcx.hpp
@@ -141,7 +141,7 @@ class baseHostCX : public Host
 	
 	
 	ssize_t processed_bytes_; //!< number of bytes processed by last process()
-	unsigned int next_read_limit_;  // limit next read() operation to this number. Zero means no restrictions.
+	ssize_t next_read_limit_;  // limit next read() operation to this number. Zero means no restrictions.
 	
 	/*! 
 	 ! If you are not attempting to do something really special, you want it to keep it as true (default). See [HostCX::auto_finish()](@ref HostCX::auto_finish) */
